* Create CODEOWNERS

* Update export tests the match new QC behavior (#225)

* Update CODEOWNERS

* SampleSet Parent Column Update for Import (#211)

Issue 37219: Parent column from older Sample Sets no longer associates the parent
* Enable creation of aliases for import to set the parent of samples using column headers

* Issue 37268: experiment-derive.api doesn't include parents in name expression (#229)

- case-insensitive check for MaterialInputs parent column name
- don't bother joining by ',' just to split it apart again later

* Issue 37869: Propagate column metadata for additional SQL aggregates (#224)

* Issue 37826: Lookup from linked schema query to a non-linked schema target fails to apply proper ContainerFilter (#226)

Also, prompt for parameters on linked schema queries that include them

New QueryWebPart option to suppress filters and parameter values in the header

* 37756: Junit page no longer respects suite (DRT, BVT) (#222)

* Issue 37914: Select All from genotyping runs grid only selects first page of data (#236)

* Item 5900: Sample Set API updates for sample set create/update/delete (#214)

- add NameExpression column to sample set table info (default to hidden)
- add DesignSampleSetPermission to be used for various API actions and SampleSetDomainKind
- refactor CreateSampleSetAction and UpdateMaterialSourceAction bodies to work for both form view and api actions

* Issue 37861: ExpData.getWebDavURL returns double encoded path in S3 folders (#237)

* Issue 37891: SampleSet: failing to resolve sample parent throws ConversionException (#238)

- rethrow RemapPostConvert's ConversionException as ValidationException

* 5869: Consolidate 19.2 release scripts (#235)

* Fix flow tests

* Minor fixup after merge branch 'release19.2-SNAPSHOT' into fb_merge_19.2_develop

* Item #5846: prevent deletion of samples that are inputs to assays or have derived samples (#230)

* 37892: Inserting rows with empty columns via list file watcher gives … (#240)

* Issue 36599: SVG icons in lineage viewer not showing up in Firefox (#241)

- add width and height attribute to svg icons for use in vis.js graph

https://github.com/almende/vis/issues/1736
https://bugzilla.mozilla.org/show_bug.cgi?id=700533

* Issue 33909: Suppress some Messages menu content when not in Page Admin Mode (#243)

'Admin' option suppressed when not in Page Admin mode. 'Email' submenu option suppressed when not in Page Admin Mode, and replaced by 'Email Preferences.'

* 5869: Consolidate 19.2 release scripts (#244)

Missed the core module version bump (?!)

* Fix build - complete variable rename (#248)

* Support JDK 14 and module version unit test (#250)

* Nominal support for JDK 14, now available as an Early-Access Release. Seems to work fine.

* Test case that verifies release-time module bumps for modules we manage

Calculate valid product versions instead of maintaining hard-coded lists in tests

Fix UnreachableScriptsAction: check only the modules we manage, use calculated collection of product versions

@Override, other warnings, spelling

* sample insert deadlock 37518 (#249)

* Issue 37518: deadlock when concurrently inserting samples with lineage
* for Samples and DataClass Data, insert into exp.object during persist rows to avoid having to ensure an object during syncRunEdges
* Add objectid to exp.data, exp.experimentrun, exp.material

* Item 5864: Type-Dependent Layout (Domain Designer) (#255)

Make JDK JavaDoc base URL available to client code via LABKEY.jdkJavaDocLinkPrefix

Dependency for https://github.com/LabKey/glass-components/pull/35

* Prevent ConcurrentModificationException in ClrAssemblyManager (#253)

+++ api/src/org/labkey/api/data/StatementUtils.java
@@ -88,7 +88,7 @@ public class StatementUtils
     // builder style methods
     //
 
-    private StatementUtils(@NotNull Operation op, @NotNull TableInfo table)
+    public StatementUtils(@NotNull Operation op, @NotNull TableInfo table)
     {
         _operation = op;
         _dialect = table.getSqlDialect();
@@ -125,25 +125,26 @@ public class StatementUtils
         return this;
     }
 
-    private StatementUtils noupdate(Set<String> noupdate)
+    public StatementUtils noupdate(Set<String> noupdate)
     {
-        _dontUpdateColumnNames = noupdate;
+        if (null != noupdate)
+            _dontUpdateColumnNames.addAll(noupdate);
         return this;
     }
 
-    private StatementUtils updateBuiltinColumns(boolean b)
+    public StatementUtils updateBuiltinColumns(boolean b)
     {
         _updateBuiltInColumns = b;
         return this;
     }
 
-    private StatementUtils selectIds(boolean b)
+    public StatementUtils selectIds(boolean b)
     {
         _selectIds = b;
         return this;
     }
 
-    private StatementUtils allowSetAutoIncrement(boolean b)
+    public StatementUtils allowSetAutoIncrement(boolean b)
     {
         _allowUpdateAutoIncrement = b;
         return this;
@@ -154,7 +155,6 @@ public class StatementUtils
         _allowInsertByLookupDisplayValue = b;
         return this;
     }
-
 
 
     /**
@@ -374,8 +374,7 @@ public class StatementUtils
         }
     }
 
-
-    private Parameter.ParameterMap createStatement(Connection conn, @Nullable Container c, User user) throws SQLException
+    public Parameter.ParameterMap createStatement(Connection conn, @Nullable Container c, User user) throws SQLException
     {
         if (!(_table instanceof UpdateableTableInfo))
             throw new IllegalArgumentException("Table must be an UpdateableTableInfo");
@@ -467,14 +466,20 @@ public class StatementUtils
         List<? extends DomainProperty> properties = Collections.emptyList();
 
         boolean hasObjectURIColumn = objectURIColumnName != null && table.getColumn(objectURIColumnName) != null;
+        boolean alwaysInsertExpObject = hasObjectURIColumn && updatable.isAlwaysInsertExpObject();
         if (hasObjectURIColumn)
             _dontUpdateColumnNames.add(objectURIColumnName);
+// TODO Should we add created and createdby? Or make the caller decide?
+//        _dontUpdateColumnNames.add("Created");
+//        _dontUpdateColumnNames.add("CreatedBy");
 
-        if (null != domain && null != domainKind && StringUtils.isEmpty(domainKind.getStorageSchemaName()))
+        boolean isMaterializedDomain = null != domain && null != domainKind && StringUtils.isNotEmpty(domainKind.getStorageSchemaName());
+
+        if (alwaysInsertExpObject || (null != domain && !isMaterializedDomain))
         {
-            properties = domain.getProperties();
+            properties = (null==domain||isMaterializedDomain) ? Collections.emptyList() : domain.getProperties();
 
-            if (!properties.isEmpty())
+            if (alwaysInsertExpObject || !properties.isEmpty())
             {
                 if (!_dialect.isPostgreSQL() && !_dialect.isSqlServer())
                     throw new IllegalStateException("Domains are only supported for sql server and postgres");
@@ -492,27 +497,39 @@ public class StatementUtils
                 // Grab the object's ObjectId based on the pk of the base table
                 if (hasObjectURIColumn)
                 {
-                    sqlfPreselectObject.append(setKeyword).append(objectURIVar).append(" = (");
-                    sqlfPreselectObject.append("SELECT ").append(table.getColumn(objectURIColumnName).getSelectName());
-                    sqlfPreselectObject.append(" FROM ").append(table.getSelectName());
-                    sqlfPreselectObject.append(getPkWhereClause(keys));
-                    sqlfPreselectObject.append(");\n");
+                    if (Operation.merge == _operation)
+                    {
+                        // this seems overkill actually, but I'm focused on optimizing insert right now (MAB)
+                        sqlfPreselectObject.append(setKeyword).append(objectURIVar).append(" = COALESCE((");
+                        sqlfPreselectObject.append("SELECT ").append(table.getColumn(objectURIColumnName).getSelectName());
+                        sqlfPreselectObject.append(" FROM ").append(table.getSelectName());
+                        sqlfPreselectObject.append(getPkWhereClause(keys));
+                        sqlfPreselectObject.append("),");
+                        appendParameterOrVariable(sqlfPreselectObject, objecturiParameter);
+                        sqlfPreselectObject.append(");\n");
+                    }
+                    else
+                    {
+                        sqlfPreselectObject.append(setKeyword).append(objectURIVar).append(" = ");
+                        appendParameterOrVariable(sqlfPreselectObject, objecturiParameter);
+                        sqlfPreselectObject.append(";\n");
+                    }
                 }
 
                 SQLFragment sqlfWhereObjectURI = new SQLFragment();
-                sqlfWhereObjectURI.append("((").append(objectURIVar).append(" IS NOT NULL AND ObjectURI = ").append(objectURIVar).append(")");
-                sqlfWhereObjectURI.append(" OR ObjectURI = ");
-                appendParameterOrVariable(sqlfWhereObjectURI, objecturiParameter);
-                sqlfWhereObjectURI.append(")");
+                sqlfWhereObjectURI.append("(ObjectURI = ").append(objectURIVar).append(")");
 
                 // In the update case, it's still possible that there isn't a row in exp.Object - there might have been
                 // no properties in the domain when the row was originally inserted
-                sqlfInsertObject.append("INSERT INTO exp.Object (container, objecturi) ");
+                sqlfInsertObject.append("INSERT INTO exp.Object (container, objecturi, ownerobjectid) ");
                 sqlfInsertObject.append("SELECT ");
                 appendParameterOrVariable(sqlfInsertObject, containerParameter);
                 sqlfInsertObject.append(" AS Container,");
                 appendParameterOrVariable(sqlfInsertObject, objecturiParameter);
-                sqlfInsertObject.append(" AS ObjectURI WHERE NOT EXISTS (SELECT ObjectURI FROM exp.Object WHERE Container = ");
+                sqlfInsertObject.append(" AS ObjectURI, ");
+                Integer ownerObjectId = updatable.getOwnerObjectId();
+                sqlfInsertObject.append( null == ownerObjectId ? "NULL" : String.valueOf(ownerObjectId) ).append(" AS OwnerObjectId");
+                sqlfInsertObject.append(" WHERE NOT EXISTS (SELECT ObjectURI FROM exp.Object WHERE Container = ");
                 appendParameterOrVariable(sqlfInsertObject, containerParameter);
                 sqlfInsertObject.append(" AND ").append(sqlfWhereObjectURI).append(");\n");
 
@@ -522,7 +539,7 @@ public class StatementUtils
                 appendParameterOrVariable(sqlfSelectObject, containerParameter);
                 sqlfSelectObject.append(" AND ").append(sqlfWhereObjectURI).append(");\n");
 
-                if (Operation.insert != _operation)
+                if (Operation.insert != _operation && !properties.isEmpty())
                 {
                     // Clear out any existing property values for this domain
                     sqlfDelete.append("DELETE FROM exp.ObjectProperty WHERE ObjectId = ");
